backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
cryptography==42.0.0
